requirements.txt
flask==2.3.3
opencv-python==4.11.0.86
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
